FIX: export RepositoryConfig UPDATE: change serverConfig struct location
